Added even more measures to stop resque from randomly decided to remove or not to remove processes

+++ lib/tasks/resque.rake
@@ -46,9 +46,11 @@ namespace :resque do
       pids = Resque.workers[0].worker_pids.join(" ")
       Resque.workers.each {|w| w.unregister_worker}
 
-      syscmd = "kill -s QUIT #{pids}"
-      puts "Running syscmd: #{syscmd}"
-      puts `#{syscmd} > /dev/null 2>&1`
+      if Resque.workers.length > 0
+        syscmd = "kill -s QUIT #{pids}"
+        puts "Running syscmd: #{syscmd}"
+        puts `#{syscmd} > /dev/null 2>&1`
+      end
     else
       puts "No workers to kill"
     end
